Add shot-parsing helpers
