+++ pooltool/ani/modes/ball_in_hand.py
@@ -14,8 +14,8 @@ from pooltool.ani.globals import Global
 from pooltool.ani.menu import TextOverlay
 from pooltool.ani.modes.datatypes import BaseMode, Mode
 from pooltool.ani.mouse import MouseMode, mouse
+from pooltool.ani.scene import visual
 from pooltool.ruleset.datatypes import BallInHandOptions
-from pooltool.system.render import visual
 from pooltool.utils import panda_path
 
 FONT_OPACITY = 0.95
