+++ pooltool/ani/modes/pick_ball.py
@@ -12,8 +12,8 @@ from pooltool.ani.camera import cam
 from pooltool.ani.globals import Global
 from pooltool.ani.modes.datatypes import BaseMode, Mode
 from pooltool.ani.mouse import MouseMode, mouse
+from pooltool.ani.scene import visual
 from pooltool.system.datatypes import multisystem
-from pooltool.system.render import visual
 
 
 class PickBallMode(BaseMode):
