+++ pooltool/ani/modes/shot.py
@@ -7,12 +7,11 @@ from pooltool.ani.action import Action
 from pooltool.ani.camera import cam
 from pooltool.ani.collision import cue_avoid
 from pooltool.ani.globals import Global
-from pooltool.ani.hud import hud
 from pooltool.ani.modes.datatypes import BaseMode, Mode
 from pooltool.ani.mouse import MouseMode, mouse
+from pooltool.ani.scene import PlaybackMode, visual
 from pooltool.objects.ball.datatypes import BallHistory
 from pooltool.system.datatypes import multisystem
-from pooltool.system.render import PlaybackMode, visual
 
 
 class ShotMode(BaseMode):
@@ -60,8 +59,7 @@ class ShotMode(BaseMode):
         if playback_mode is not None:
             visual.animate(playback_mode)
 
-        system_cue = multisystem.active.cue
-        hud.update_cue(system_cue, multisystem.active.balls[system_cue.cue_ball_id])
+        visual.update_hud()
 
         tasks.register_event("space", visual.toggle_pause)
         tasks.register_event("arrow_up", visual.speed_up)
@@ -105,16 +103,15 @@ class ShotMode(BaseMode):
         tasks.add(self.shared_task, "shared_task")
 
     def exit(self, key="soft"):
-        """Exit shot mode
+        """Exit shot mode.
 
-        Parameters
-        ==========
-        key : str, 'soft'
-            Specifies how shot mode should be exited. Can be any of {'advance', 'reset',
-            'soft'}. 'advance' and 'reset' end the animation, whereas 'soft' exits shot
-            mode with the animations still playing. 'advance' sets the system state to
-            the end state of the shot, whereas 'reset' returns the system state to the
-            start state of the shot.
+        Args:
+            key:
+                Specifies how shot mode should be exited. Can be any of {'advance',
+                'reset', 'soft'}. 'advance' and 'reset' end the animation, whereas
+                'soft' exits shot mode with the animations still playing. 'advance' sets
+                the system state to the end state of the shot, whereas 'reset' returns
+                the system state to the start state of the shot.
         """
         assert key in {"advance", "reset", "soft"}
 
@@ -125,63 +122,55 @@ class ShotMode(BaseMode):
 
         if key == "advance":
             # New shot means new system. Depending how ShotMode was entered, it may
-            # already exist in multisystem. If the most recently added system is
-            # unsimulated, it already exists. Otherwise, it needs to be created, using
-            # most recent system as template.
-            if multisystem[-1].simulated:
+            # already exist in the multisystem. If the most recently added system is
+            # unsimulated, it is considered the new system. Otherwise, it needs to be
+            # created, using most recent system as template.
+            new_system_exists = multisystem[-1].simulated
+
+            if new_system_exists:
                 # The shot is processed and advanced now, because (1) the shot animation
                 # has ended or the user has requested to take the next shot and (2) it
                 # hasn't been processed yet, since if it had, the latest system would be
                 # unsimulated.
                 Global.game.process_and_advance(multisystem[-1])
                 if Global.game.shot_info.game_over:
-                    # The game is over, so scrap that, that, head to game over screen
                     Global.mode_mgr.change_mode(Mode.game_over)
 
                 if multisystem.active_index != multisystem.max_index:
                     # The currently rendered shot isn't the most recent shot, and it's
-                    # the most recent shot we want to advance from. So render it.
-                    multisystem.set_active(-1)
-                    visual.attach_system(multisystem.active)
-                    visual.buildup()
+                    # the most recent shot we want to advance from. So we render it as
+                    # an intermediate step.
+                    visual.switch_rendered_system(multisystem_idx=-1)
 
                     # self.quats is a vestige held in BallRender. We need to calculate
-                    # it so we know the final orientation of each ball. We don't have it
-                    # because visual.buildup() was just called.
+                    # it so we know the final orientation of each ball.
                     for ball_render in visual.balls.values():
                         ball_render.set_quats(ball_render._ball.history_cts)
 
-                # Copy the shot
-                new = multisystem.active.copy()
+                new_system = multisystem.active.copy()
 
-                for ball, old_render in zip(new.balls.values(), visual.balls.values()):
+                for ball, old_render in zip(
+                    new_system.balls.values(), visual.balls.values()
+                ):
                     # Set the initial ball orientations of the new shot to match the
                     # final ball orientations of the old shot
                     ball.initial_orientation = old_render.get_final_orientation()
 
-                new.reset_history()
-                multisystem.append(new)
+                new_system.reset_history()
+                multisystem.append(new_system)
 
             # Switch shots
-            multisystem.set_active(-1)
-            visual.attach_system(multisystem.active)
-            visual.buildup()
+            visual.switch_rendered_system(multisystem_idx=-1)
 
             cue_avoid.init_collisions()
             multisystem.active.cue.reset_state()
             visual.cue.set_render_state_as_object_state()
 
-            # Set the HUD
-            system_cue = multisystem.active.cue
-            hud.update_cue(system_cue, multisystem.active.balls[system_cue.cue_ball_id])
-
         elif key == "reset":
             if multisystem.active_index != len(multisystem) - 1:
                 # Replaying shot that is not most recent. Teardown and then buildup most
                 # recent
-                multisystem.set_active(-1)
-                visual.attach_system(multisystem.active)
-                visual.buildup()
+                visual.switch_rendered_system(multisystem_idx=-1)
                 cue_avoid.init_collisions()
             else:
                 visual.reset_animation()
@@ -253,106 +242,61 @@ class ShotMode(BaseMode):
                 enter_kwargs=dict(load_prev_cam=True),
             )
 
-        elif (
-            self.keymap[Action.parallel] and self.view_only
-        ):  # Only process in ShotViewer mode
+        elif self.keymap[Action.parallel] and self.view_only:
             self.keymap[Action.parallel] = False
             if visual.is_parallel_mode:
-                # Exit parallel mode
                 visual.exit_parallel_mode()
-                self.change_animation(multisystem.active_index)
             else:
-                # First pause and reset any current animation
-                was_playing = not visual.paused
-                visual.pause_animation()
-                visual.reset_animation()
-
-                # Enter parallel mode
                 visual.setup_parallel_mode()
-
-                # Start the animation if it was previously playing
-                if was_playing:
-                    visual.animate(PlaybackMode.LOOP)
-                else:
-                    # Just build the animation but don't play it
-                    visual.paused = True
 
         elif self.keymap[Action.prev_shot]:
             self.keymap[Action.prev_shot] = False
-            shot_index = multisystem.active_index - 1
-            while True:
-                if shot_index < 0:
-                    shot_index = len(multisystem) - 1
-                if (
-                    len(multisystem[shot_index].events)
-                    or shot_index != len(multisystem) - 1
-                ):
-                    break
-                shot_index -= 1
-            self.change_animation(shot_index)
+            shot_index = self._find_previous_valid_shot()
+            visual.switch_to_shot(shot_index)
+            cue_avoid.init_collisions()
 
         elif self.keymap[Action.next_shot]:
             self.keymap[Action.next_shot] = False
-            shot_index = multisystem.active_index + 1
-            while True:
-                if shot_index == len(multisystem):
-                    shot_index = 0
-                if (
-                    len(multisystem[shot_index].events)
-                    or shot_index != len(multisystem) - 1
-                ):
-                    break
-                shot_index += 1
-            self.change_animation(shot_index)
+            shot_index = self._find_next_valid_shot()
+            visual.switch_to_shot(shot_index)
+            cue_avoid.init_collisions()
 
         return task.cont
 
-    @staticmethod
-    def change_animation(shot_index):
-        """Switch to a different system in the system collection"""
-        # Was the animation playing before?
-        was_playing = not visual.paused
+    def _find_previous_valid_shot(self) -> int:
+        """Find the previous valid shot, wrapping around if needed."""
+        shot_index = multisystem.active_index - 1
 
-        # Pause animation during transition to prevent visual glitches
-        visual.pause_animation()
+        # Search backwards through shots, wrapping to end if needed
+        for _ in range(len(multisystem)):
+            if shot_index < 0:
+                shot_index = len(multisystem) - 1
 
-        # Switch shots
-        multisystem.set_active(shot_index)
+            if self._is_valid_shot(shot_index):
+                return shot_index
 
-        if visual.is_parallel_mode:
-            # In parallel mode, just update which system is active
-            visual.update_parallel_active(shot_index)
+            shot_index -= 1
 
-            # Set the HUD
-            system_cue = multisystem.active.cue
-            hud.update_cue(system_cue, multisystem.active.balls[system_cue.cue_ball_id])
+        # Fallback to current shot if no valid shot found
+        return multisystem.active_index
 
-            # Resume animation if it was playing
-            if was_playing:
-                visual.resume_animation()
+    def _find_next_valid_shot(self) -> int:
+        """Find the next valid shot, wrapping around if needed."""
+        shot_index = multisystem.active_index + 1
 
-            return
+        # Search forwards through shots, wrapping to start if needed
+        for _ in range(len(multisystem)):
+            if shot_index >= len(multisystem):
+                shot_index = 0
 
-        # Standard mode - clear any existing animation
-        visual.reset_animation()
+            if self._is_valid_shot(shot_index):
+                return shot_index
 
-        # Setup new system
-        visual.attach_system(multisystem.active)
-        visual.buildup()
+            shot_index += 1
 
-        # Initialize the animation
-        visual.build_shot_animation()
+        # Fallback to current shot if no valid shot found
+        return multisystem.active_index
 
-        # Changing to a different shot is considered advanced maneuvering, so we enter
-        # loop mode if it was previously playing
-        if was_playing:
-            visual.animate(PlaybackMode.LOOP)
-        else:
-            # Ready to play but maintain paused state
-            visual.playback(PlaybackMode.LOOP)
-
-        cue_avoid.init_collisions()
-
-        # Set the HUD
-        system_cue = multisystem.active.cue
-        hud.update_cue(system_cue, multisystem.active.balls[system_cue.cue_ball_id])
+    def _is_valid_shot(self, shot_index: int) -> bool:
+        """Check if a shot index represents a valid shot to navigate to."""
+        return multisystem[shot_index].simulated or shot_index != len(multisystem) - 1
