+++ pooltool/system/__init__.py
@@ -1,13 +1,11 @@
 """The system container and its associated objects"""
 
 from pooltool.system.datatypes import MultiSystem, System, multisystem
-from pooltool.system.render import SystemController, SystemRender, visual
+from pooltool.system.render import SystemRender
 
 __all__ = [
     "System",
     "MultiSystem",
     "multisystem",
     "SystemRender",
-    "SystemController",
-    "visual",
 ]
