+++ pooltool/utils/strenum.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from enum import Enum, auto
 from typing import Any, TypeVar
 
@@ -42,3 +44,7 @@ class StrEnum(str, Enum):
         Return the lower-cased version of the member name.
         """
         return name.lower()
+
+    @classmethod
+    def members_as_list(cls: type[_S]) -> list[_S]:
+        return list(cls.__members__.values())
